[FIX] sale_commission: Salesmen can see invoice from SO
Steps to reproduce:

- With a user with only sales permission and no invoicing one.
- Go to an invoiced sales order.
- Click on the smart-button "Invoice".

You get a permission error on `account.invoice.line.agent` model.

With this commit, we add read permission to salesmen over that object
to avoid the problem. The other option would be to protect the view
under the invoicing group, but it's interesting that salesmen can
check its own invoices commissions.

TT44083